### 2022-07-03 #4543 @Kevin-Mattheus-Moerman
@AldoGl the paper compilation error says: `Problem with ORCID (0000-0000-0000-0000) for Marco Benedetti. Invalid ORCID`, can you please check your ORCID entries to make sure they are valid? You can remove the ORCID field if a person does not have an ORCID.
Once you've made changes you can call the following here to update the draft paper: `@editorialbot generate pdf`
