### 2022-07-04 ef0cedd
joss-paper: small fixes
