### 2022-07-04 af1affb
joss-paper: add doi to bib references
fix
