+++ paper/paper.bib
@@ -1,11 +1,3 @@
-@ARTICLE{Fortunato2010,
-  author  = {Fortunato, S.},
-  title   = {Community detection in graphs},
-  journal = {Phys. Rep.-Rev. Sec. Phys. Lett.},
-  volume  = {486},
-  year    = {2010},
-  pages   = {75-174}
-}
 
 @article{turrell2016agent,
   title={Agent-based models: understanding the economy from the bottom up},
@@ -45,7 +37,8 @@
   number={2},
   pages={143--186},
   year={1971},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/0022250X.1971.9989794}
 }
 
 @article{simoes2008stochastic,
@@ -56,7 +49,8 @@
   number={22},
   pages={555--566},
   year={2008},
-  publisher={The Royal Society London}
+  publisher={The Royal Society London},
+  doi={10.1098/rsif.2007.1206}
 }
 
 
@@ -68,14 +62,16 @@
   number={7},
   pages={e1009146},
   year={2021},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1009146}
 }
 
 @article{baptista2016macroprudential,
   title={Macroprudential policy in an agent-based model of the UK housing market},
   author={Baptista, Rafa and Farmer, J Doyne and Hinterschweiger, Marc and Low, Katie and Tang, Daniel and Uluc, Arzu},
   year={2016},
-  publisher={Bank of England Working Paper}
+  publisher={Bank of England Working Paper},
+  doi={10.2139/ssrn.2850414}
 }
 
 @article{catapano2021macroprudential,
@@ -83,7 +79,8 @@
   author={Catapano, Gennaro and Franceschi, Francesco and Loberto, Michele and Michelangeli, Valentina},
   journal={Bank of Italy Temi di Discussione (Working Paper) No},
   volume={1338},
-  year={2021}
+  year={2021},
+  doi={10.2139/ssrn.3891583}
 }
 
 @article{richiardi2012agent,
@@ -110,7 +107,8 @@
   number={2},
   pages={266--294},
   year={1997},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/264029.264064}
 }
 
 @inproceedings{tisue2004netlogo,
@@ -135,7 +133,8 @@
   author={Bajer, Luk{\'a}{\v{s}} and Pitra, Zbyn{\v{e}}k and Hole{\v{n}}a, Martin},
   booktitle={Proceedings of the Companion Publication of the 2015 Annual Conference on Genetic and Evolutionary Computation},
   pages={1143--1150},
-  year={2015}
+  year={2015},
+  doi={10.1145/2739482.2768468}
 }
 
 @article{ward2016dynamic,
@@ -146,14 +145,16 @@
   number={4},
   pages={150703},
   year={2016},
-  publisher={The Royal Society}
+  publisher={The Royal Society},
+  doi={10.1098/rsos.150703}
 }
 
 @article{knysh2016blackbox,
   title={Blackbox: A procedure for parallel optimization of expensive black-box functions},
   author={Knysh, Paul and Korkolis, Yannis},
   journal={arXiv preprint arXiv:1605.00998},
-  year={2016}
+  year={2016},
+  doi={10.48550/arXiv.1605.00998}
 }
 
 @article{grazzini2017bayesian,
@@ -163,7 +164,8 @@
   volume={77},
   pages={26--47},
   year={2017},
[…]
