### 2022-07-11 #4543 @AldoGl
> @AldoGl would you be able to take a look at [this list of people](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit?usp=sharing) and suggest a few that you feel would be a good fit to review this work? (Please don't `@` them, just let me know their handles :))

Dear @drvinceknight,

I had a look at the list you sent me. I found the following people who are experienced in Python and who work on computational economics and simulation methods. They could be particularly interested and suited to review our work, and the highlighted names seem the best candidates

- **rickecon**
- **igarizio**
- Athene-ai
- seyhunsaral

Alternatively, I also identified people experienced in Python and working on simulation methods and agent-based modelling in other social fields (not in economics). These might also be interested in reviewing our work as our tool is not intended solely for economic ABMs. Also in this case the highlighted names indicate the candidates who seem the most appropriate

- **chrisvoncsefalvay**
- **Christovis**
- **dinacmistry; DMistry-IDM**
- rusu24edward
- jamesdamillington
- RafalKucharskiPK

I hope this helps.

Best regards,
Aldo
