### 2022-07-26 #4543 @drvinceknight
@rickecon @igarizio @Athene-ai @seyhunsaral I'm pinging you to ask if you would be able to assist with reviewing this submissions to JOSS (The Journal of Open Source Software).

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: “Black-it: A Ready-to-Use and Easy-to-Extend Calibration Kit for Agent-based Models”. You can find more information at the top of this Github issue.
