### 2022-07-29 #4622 @editorialbot
[REVIEW]: Black-it: A Ready-to-Use and Easy-to-Extend Calibration Kit for Agent-based Models
**Submitting author:** @AldoGL (<a href="http://orcid.org/0000-0002-4737-2878">Aldo Glielmo</a>)
**Repository:** https://github.com/bancaditalia/black-it/
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v0.1.2
