### 2022-08-12 #4622 @Athene-ai
I have just finished my review and here below are some additional comments.

Additional comments:

_Contribution and authorship_: Seem that in the contribution description is missing

-Installation_: The reviewer suggests expanding the installation description especially if the package will be installed by a newbie

_Example usage_: This part should be implemented

_Functionality documentation_ and _Automated tests_: these parts should be better implemented
