### 2022-08-29 940297b
fixed typo in AUTHORS.md
