### 2022-08-29 473195d
docs: split further_info into simulator_interface and checkpointing_parallelisation
