+++ docs/examples.md
@@ -30,7 +30,7 @@ This is an agent-based [SIR](https://en.wikipedia.org/wiki/Compartmental_models_
 
 In this model, each agent has a probability of being infected by having a contact with another agent and a probability of recovering. Such probabilities are modeled by the SIR parameters, while the contact network is modeled by a random graph generated by a [Watts-Strogatz model](https://en.wikipedia.org/wiki/Watts%E2%80%93Strogatz_model).
 
-As you will notice, in this tutorial, the model simulation is not implemented directly in Python but an external simulator is called and run on a [docker container](https://docs.docker.com/get-docker/). For more information on how to use this option, check [how to interface with external simulators](furtherinfo.md#how-to-interface-with-external-simulators).
+As you will notice, in this tutorial, the model simulation is not implemented directly in Python but an external simulator is called and run on a [docker container](https://docs.docker.com/get-docker/). For more information on how to use this option, check the [simulator interface](simulator_interface.md) page.
 
 The chosen loss function is the quadratic loss and the chosen samplers are Halton, random forest and best batch samplers.
 
