### 2022-08-29 6ad35b1
docs: split further_info into simulator_interface and checkpointing_parallelisation
