### 2022-08-29 e3e3bfe
readme: improve description in the Docs section
