### 2022-08-29 6f6240d
docs: add a graphic illustration to description.md
