### 2022-08-29 ab92240
docs: improve description in simulator_interface.md
