### 2022-08-29 fdb91ae
readme: improve description of the quick example
