### 2022-08-29 83200ab
test: skip 'test_gaussian_process_2d' due to suspected flakiness
This change is a response to the main PR comment in https://github.com/bancaditalia/black-it/pull/9.
We leave the investigation of why the involved test does not behave robustly as a future work.
