### 2022-08-29 69173cf
static: fix mypy error in black_it/search_space.py
The error was the following:

    black_it/search_space.py:79: error: Argument 1 to "append" of "list" has incompatible type "ndarray[Any, dtype[floating[_16Bit]]]"; expected "ndarray[Any, dtype[floating[_64Bit]]]"
