### 2022-09-01 dcd061d
ci: run docs and main workflow on push to main
