+++ .github/workflows/test.yml
@@ -1,0 +1,53 @@
+name: test
+
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+
+jobs:
+  run:
+    continue-on-error: True
+    runs-on: ${{ matrix.os }}
+
+    strategy:
+      matrix:
+        os: [ubuntu-latest]
+        python-version: [3.8]
+
+    timeout-minutes: 30
+
+    steps:
+    - uses: actions/checkout@master
+    - uses: actions/setup-python@master
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run:  pip install tox poetry
+#    # set up environment depending on the platform in use
+#    - if: matrix.os == 'ubuntu-latest'
+#      name: Install dependencies (ubuntu-latest)
+#      run: ...
+#    - if: matrix.os == 'macos-latest'
+#      name: Install dependencies (macos-latest)
+#      run: ...
+#    - if: matrix.os == 'windows-latest'
+#      name: Install dependencies (windows-latest)
+#      env:
+#        ACTIONS_ALLOW_UNSECURE_COMMANDS: true
+#      run: ...
+    - name: Unit tests and coverage
+      run: |
+        tox -e py${{ matrix.python-version }} -- -m 'not e2e'
+#    - name: Upload coverage to Codecov
+#      uses: codecov/codecov-action@v1
+#      with:
+#        token: ${{ secrets.CODECOV_TOKEN }}
+#        file: ./coverage.xml
+#        flags: unittests
+#        name: codecov-umbrella
+#        fail_ci_if_error: true
+    - name: Notebook tests
+      run: |
+        tox -e py${{ matrix.python_version }}-nb
