### 2022-09-08 ed170c5
ci: split main.yml into lint and test, and removed checks for changed files in lint
the removal is needed to allow the same CIs to work both for pull requests and for merges on the main branch
