### 2022-09-09 ea4bd71
lint: ignore safety checks for ipywidgets until new package requirements update
