+++ tox.ini
@@ -108,7 +108,7 @@ commands = bandit --configfile .bandit.yaml --recursive black_it tests scripts e
 skipsdist = False
 skip_install = False
 deps = safety>=1.10.3,<1.11.0
-commands = safety check -i 44715 -i 44716 -i 44717
+commands = safety check -i 44715 -i 44716 -i 44717 -i 50664 -i 50463
 
 [testenv:vulture]
 skipsdist = True
