### 2022-09-09 f5268d0
chore: update ipywidgets version in pyproject.toml to ^8.0.0.
