+++ pyproject.toml
@@ -35,7 +35,7 @@ classifiers = [
 [tool.poetry.dependencies]
 python = ">=3.8,<3.11"
 GPy = "^1.10.0"
-ipywidgets = "^7.7.0"
+ipywidgets = "^8.0.0"
 matplotlib = "^3.5.2"
 numpy = "^1.22.3"
 pandas = "^1.4.2"
