### 2022-09-09 87a61ab
CIs: extend unit testing to python [3.8, 3.9, 3.10]
fix


fix
