+++ .github/workflows/test.yml
@@ -14,7 +14,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest]
-        python-version: [3.8]
+        python-version: [3.8, 3.9, "3.10"]
 
     timeout-minutes: 30
 
