### 2022-09-09 e54b370
ci: use tox to generate documentation
fix
