### 2022-09-17 4cf2be1
docs: changed html specification of figure in description.md
