### 2022-09-17 43fef2f
joss-paper: add new paragraph in Statement of need
