### 2022-09-17 8221916
docs: quickfix of figure path in description.md
