### 2022-09-17 #4622 @AldoGl
Thank you @aseyq for taking the time to learn about our work and for assessing our package and manuscript.

Following your suggestions we did the following.

- To make the documentation more easily visible, we linked the documentation website in the right column of the GitHub page (in the ”About” section), as it is often done for GitHub packages.

- The documentation website does contain an API description section (see e.g., https://bancaditalia.github.io/black-it/calibrator/) To make this more visible we linked some of its pages directly within the README page, under the Docs section, where we also provide an essential description of the main abstractions of the package.

- We improved the documentation by adding a clearer explanation of how the package works in the new section https://bancaditalia.github.io/black-it/description/, also adding a simple illustration.

- We improved the documentation by adding a specific section describing how Black-it can be used with standard models https://bancaditalia.github.io/black-it/simulator_interface/. This was previously described in a section named “Further reading” and partially illustrated in the Jupyter notebooks, but we agree with the referee that a clearer explanation was beneficial. Since Black-it is indeed language agnostic, we mention this as suggested by the reviewer.

- We added a new paragraph to our manuscript where we differentiate our ABM calibration software from standard function optimisation libraries, highlighting the defining features and strengths of Black-it.
