### 2022-09-19 #4622 @aseyq
Dear @AldoGl,
Thank you very much for the work done. I think this is a very nice piece of software and I am sure the community will benefit from it. I am looking forward to seeing it published.

Just let me write my comments:
- Thank you for adding the paragraph. Similarly, thank you for the improvements in the documentation. I believe it will help people to understand the nature of the package and help them take it up.

- Of course, I was already aware of the documentation and went through both the "Further Reading" section and the API description, tested the software, and even tried to connect it with a model of my own by writing a wrapper for using it with the software. So my previous comments were not missing any of those; on the contrary the **Functionality Documentation** was specially directed to those pages. As I wrote earlier,  I think they are adequate for a designated user to navigate. Especially in terms of coverage, it can be tagged "Good" for the journal criteria. But I still think there is a big room for improvement. I believe this is especially important because the ABM community is diverse regarding their main fields and technical abilities. I certainly believe that it would be beneficial for the authors (or community members) to make this part of the documentation a bit more friendly, even risking being pedantic. However, I don't want it to be an obstacle to the publication of the paper and completing the checklist. I hope the authors will add it to their checklists and I leave it up to the second reviewer.
