### 2022-09-28 bf7c352
feat: add particle swarm sampler
Implementation of the classical particle swarm sampler.

References:

  - Shi, Yuhui, and Russell Eberhart. "A modified particle swarm optimizer."
    1998 IEEE international conference on evolutionary computation proceedings.
  - Eberhart, Russell, and James Kennedy. "A new optimizer using particle swarm theory."
    MHS'95. Proceedings of the sixth international symposium on micro machine and human science.
