+++ scripts/whitelists/package_whitelist.py
@@ -1,26 +1,25 @@
 # flake8: noqa
 # type: ignore
 # pylint: skip-file
-_.random_state  # unused attribute (black_it/samplers/best_batch.py:47)
-plot_losses_interact  # unused function (black_it/plot/plot_results.py:152)
-plot_sampling_interact  # unused function (black_it/plot/plot_results.py:218)
-plot_convergence  # unused function (black_it/plot/plot_results.py:21)
-plot_sampling  # unused function (black_it/plot/plot_results.py:81)
-plot_losses  # unused function (black_it/plot/plot_results.py:50)
-MinkowskiLoss  # unused class (black_it/loss_functions/minkowski.py:21)
-MethodOfMomentsLoss  # unused class (black_it/loss_functions/msm.py:26)
-BestBatchSampler  # unused class (black_it/samplers/best_batch.py:12)
-_.restore_from_checkpoint  # unused method (black_it/calibrator.py:118)
-ts_stats  # unused function (black_it/plot/plot_descriptive_statistics.py:10)
-RandomForestSampler  # unused class (black_it/samplers/random_forest.py:14)
-_.calibrate  # unused method (black_it/calibrator.py:229)
-GslDivLoss  # unused class (black_it/loss_functions/gsl_div.py:57)
-GaussianProcessSampler  # unused class (black_it/samplers/gaussian_process.py:15)
-Calibrator  # unused class (black_it/calibrator.py:23)
-MEAN  # unused variable (black_it/samplers/gaussian_process.py:36)
-get_mom_ts  # unused function (black-it/black_it/utils/time_series.py)
-FourierLoss  # unused class (black_it/loss_functions/fourier.py:26)
-ideal_low_pass_filter  # unused function (black-it/black_it/loss_functions/fourier.py:26)
-set_samplers  # unused method (black-it/black_it/calibrator.py:153)
+_.set_samplers  # unused method (black_it/calibrator.py:180)
+_.restore_from_checkpoint  # unused method (black_it/calibrator.py:203)
+_.calibrate  # unused method (black_it/calibrator.py:307)
+ideal_low_pass_filter  # unused function (black_it/loss_functions/fourier.py:35)
+FourierLoss  # unused class (black_it/loss_functions/fourier.py:86)
+GslDivLoss  # unused class (black_it/loss_functions/gsl_div.py:66)
+MinkowskiLoss  # unused class (black_it/loss_functions/minkowski.py:27)
+ts_stats  # unused function (black_it/plot/plot_descriptive_statistics.py:26)
+plot_convergence  # unused function (black_it/plot/plot_results.py:75)
+plot_losses  # unused function (black_it/plot/plot_results.py:111)
+plot_sampling  # unused function (black_it/plot/plot_results.py:140)
+plot_losses_interact  # unused function (black_it/plot/plot_results.py:212)
+plot_sampling_interact  # unused function (black_it/plot/plot_results.py:287)
+BestBatchSampler  # unused class (black_it/samplers/best_batch.py:27)
+MEAN  # unused variable (black_it/samplers/gaussian_process.py:37)
+GaussianProcessSampler  # unused class (black_it/samplers/gaussian_process.py:45)
 HaltonSampler  # unused class (black_it/samplers/halton.py:33)
-RSequenceSampler  # unused class (black_it/samplers/r_sequence.py:30)
+ParticleSwarmSampler  # unused class (black_it/samplers/particle_swarm.py:28)
+_.reset  # unused method (black_it/samplers/particle_swarm.py:137)
+RSequenceSampler  # unused class (black_it/samplers/r_sequence.py:32)
+RandomForestSampler  # unused class (black_it/samplers/random_forest.py:30)
+get_mom_ts  # unused function (black_it/utils/time_series.py:25)
