### 2022-09-28 4837790
fix: harmonize pyproject.toml and tox.ini
There was a mismatch between the dependency versions installed by the test envirnoments
(specified by tox.ini) and the production/development environment (specified
by pyproject.toml).

For example:
- in pyproject.toml, pytest's version specifier was >=7.1.2,<7.2.0, but
  in tox.ini it was pytest>=6.2.5,<6.3.0
- in pyproject.toml, numpy's version specifier was ^1.22.3 (which installed 1.23.3), but
  in tox.ini it was numpy>=1.21.0,<1.22.0

These changes might have introduced errors in tests and/or linting checks,
which will be fixed in the following co
[…]
