+++ tox.ini
@@ -12,10 +12,10 @@ basepython = python3
 setenv =
     PYTHONPATH = {toxinidir}
 deps =
-    pytest>=6.2.5,<6.3.0
+    pytest>=7.1.2,<7.2.0
     pytest-cov>=3.0.0,<3.1.0
-    pytest-randomly>=3.11.0,<3.12.0
-    hypothesis[numpy]>=6.37.1,<6.38.0
+    pytest-randomly>=3.12.0,<3.13.0
+    hypothesis[numpy]>=6.46.5,<6.47.0
     ; Other test dependencies
     ; ...
     mistletoe>=0.8.2,<0.9.0
@@ -41,9 +41,9 @@ basepython = python3.10
 [testenv:py3.8-nb]
 basepython = python3.8
 deps =
-    pytest>=6.2.5,<6.3.0
-    mesa>=0.8.9,<0.9.0
-    nbmake>=1.2,<1.3
+    pytest>=7.1.2,<7.2.0
+    mesa>=0.9.0,<0.10.0
+    nbmake>=1.3.0,<1.4.0
 commands:
     pytest examples/tests_on_toy_model.ipynb --nbmake --nbmake-timeout=300
 
@@ -51,7 +51,7 @@ commands:
 skip_install = True
 deps =
     flake8>=4.0.1,<4.1.0
-    flake8-bugbear>=22.1.11,<22.2.0
+    flake8-bugbear>=22.4.25,<22.5.0
     flake8-docstrings>=1.6.0,<1.7.0
     flake8-eradicate>=1.2.0,<1.3.0
     flake8-isort>=4.1.1,<4.2.0
@@ -63,18 +63,18 @@ commands =
 skip_install = True
 deps =
     mypy>=0.931,<0.940
-    numpy>=1.21.0,<1.22.0
-    hypothesis[numpy]>=6.37.1,<6.38.0
+    numpy>=1.23.3,<1.24.0
+    hypothesis[numpy]>=6.46.5,<6.47.0
 commands =
     mypy black_it tests scripts examples
 
 [testenv:pylint]
 skipdist = True
 deps =
-    pylint>=2.12.2,<2.13.0
-    pytest>=6.2.5,<6.3.0
-    hypothesis[numpy]>=6.37.1,<6.38.0
-    mesa>=0.8.9,<0.9.0
+    pylint>=2.13.9,<2.14.0
+    pytest>=7.1.2,<7.2.0
+    hypothesis[numpy]>=6.46.5,<6.47.0
+    mesa>=0.9.0,<0.10.0
     mistletoe>=0.8.2,<0.9.0
 commands = pylint black_it tests scripts examples
 
@@ -101,7 +101,7 @@ commands = isort --check-only black_it tests scripts examples
 [testenv:bandit]
 skipsdist = True
 skip_install = True
-deps = bandit>=1.7.2,<1.8.0
+deps = bandit>=1.7.4,<1.8.0
 commands = bandit --configfile .bandit.yaml --recursive black_it tests scripts examples
 
 [testenv:safety]
@@ -113,7 +113,7 @@ commands = safety check -i 44715 -i 44716 -i 44717 -i 50664 -i 50463
 [testenv:vulture]
 skipsdist = True
 skip_install = True
-deps = vulture>=2.3,<2.4
+deps = vulture>=2.3,<3.0
 commands =
     vulture black_it scripts/whitelists/package_whitelist.py
     vulture examples scripts/whitelists/examples_whitelist.py
