+++ pyproject.toml
@@ -37,7 +37,7 @@ python = ">=3.8,<3.11"
 GPy = "^1.10.0"
 ipywidgets = "^7.7.0"
 matplotlib = "^3.5.2"
-numpy = "^1.22.3"
+numpy = ">=1.23.3,<1.24.0"
 pandas = "^1.4.2"
 scikit-learn = "^1.1.0"
 seaborn = "^0.11.2"
@@ -57,21 +57,23 @@ hypothesis = "^6.46.5"
 ipython = "^8.3.0"
 isort = "^5.10.1"
 jupyter = "^1.0.0"
-Markdown = "^3.3.7"
-markdown-include = "^0.6.0"
+Markdown = "==3.3.7"
+markdown-include = "==0.6.0"
 Mesa = "^0.9.0"
 mistletoe = "^0.8.2"
-mkdocs = "^1.3.0"
-mkdocs-autorefs = "^0.4.1"
-mkdocs-material = "^8.2.15"
-mkdocstrings = "^0.18.1"
-mknotebooks = "^0.7.1"
-mypy = "^0.950"
+mkdocs = "==1.3.0"
+mkdocs-autorefs = "==0.4.1"
+mkdocs-material = "==8.2.15"
+mkdocs-material-extensions = "==1.0.3"
+mkdocstrings = "==0.18.1"
+mkdocstrings-python-legacy = "==0.2.2"
+mknotebooks = "==0.7.1.post0.dev98"
+mypy = "^0.931"
 nbmake = "^1.3.0"
 ndlib = "^5.1.1"
 pydocstyle = "^6.1.1"
 pylint = "^2.13.9"
-pymdown-extensions = "^9.4"
+pymdown-extensions = "==9.4"
 pytest = "^7.1.2"
 pytest-cov = "^3.0.0"
 pytest-randomly = "^3.12.0"
