### 2022-09-28 027d75b
static: fix mypy errors after numpy upgrade
Fix mypy errors raised after the upgrade to numpy 1.23.3, which became
stricter in the checking of the type of elements of the NDArray generic.
