### 2022-09-28 7c626c8
lint: fix pylint errors after upgrade
- fix new error "consider-using-generator"
- add 'pylint.extensions.no_self_use' in load-plugins option due to changes in pylint 2.14
