### 2022-09-28 c07b097
chore: bump mypy to latest version 0.981
Update pyproject.toml, tox.ini and poetry.lock accordingly.
