+++ tox.ini
@@ -62,7 +62,7 @@ commands =
 [testenv:mypy]
 skip_install = True
 deps =
-    mypy>=0.931,<0.940
+    mypy>=0.981,<0.990
     numpy>=1.23.3,<1.24.0
     hypothesis[numpy]>=6.46.5,<6.47.0
 commands =
