+++ pyproject.toml
@@ -68,7 +68,7 @@ mkdocs-material-extensions = "==1.0.3"
 mkdocstrings = "==0.18.1"
 mkdocstrings-python-legacy = "==0.2.2"
 mknotebooks = "==0.7.1.post0.dev98"
-mypy = "^0.931"
+mypy = ">=0.981,<0.990"
 nbmake = "^1.3.0"
 ndlib = "^5.1.1"
 pydocstyle = "^6.1.1"
