### 2022-09-28 487dd66
chore: bump pylint to latest version 2.15.3
