+++ pyproject.toml
@@ -72,7 +72,7 @@ mypy = ">=0.981,<0.990"
 nbmake = "^1.3.0"
 ndlib = "^5.1.1"
 pydocstyle = "^6.1.1"
-pylint = "^2.13.9"
+pylint = ">=2.15.3,<2.16.0"
 pymdown-extensions = "==9.4"
 pytest = "^7.1.2"
 pytest-cov = "^3.0.0"
