### 2022-09-28 027ae63
chore: bump pylint to latest version 2.15.3
