+++ tox.ini
@@ -71,7 +71,7 @@ commands =
 [testenv:pylint]
 skipdist = True
 deps =
-    pylint>=2.13.9,<2.14.0
+    pylint>=2.15.3,<2.16.0
     pytest>=7.1.2,<7.2.0
     hypothesis[numpy]>=6.46.5,<6.47.0
     mesa>=0.9.0,<0.10.0
