### 2022-10-14 b24a997
Delete Plot.R
