### 2022-10-14 6d74967
Delete LDP.R
