### 2022-10-17 #4622 @AldoGl
Dear @drvinceknight,

I am writing to you to enquire about the status of our manuscript, to make sure that the review process is moving forward smoothly, and to know whether we can facilitate the process in any way at this stage.

Best wishes,
Aldo
