### 2022-10-19 d7ee87e
PS sampler: remove print of position and velocities even when verbose=True
