### 2022-10-19 d56322e
particle swarm sampler: remove unused "verbose" attribute
