### 2022-10-19 ad11165
test: parametrize PS tests on the option global_minimum_across_samplers
