### 2022-10-20 46b3dc0
particle swarm: update docstring description
