### 2022-10-26 9244d57
mcs.R: add questions to original R code
