### 2022-11-02 #4622 @AldoGl
Dear professor @danielskatz,

Thank you for your message, and thank you, @drvinceknight, @aseyq and @Athene-ai for the your editorial/review work.

I made a tagged release of v0.1.2 of the software, you can find it here https://github.com/bancaditalia/black-it/tree/v0.1.2.

I archived the same release on Zenodo with the DOI 10.5281/zenodo.7273564, this is the corresponding link https://doi.org/10.5281/zenodo.7273564.

I checked that the metadata of the Zenodo archive are correct, and I included the ORCID number of the authors.

Please let me know how I can further help the publication process.
