### 2022-11-02 9632206
bump version to 0.1.2, update HISTORY.md, add ignores to safety checks
fix
