+++ HISTORY.md
@@ -1,3 +1,16 @@
-# History
+Release History
+===============
 
-TODO
+0.1.2 (2022-11-02)
+------------------
+* Dependencies: regenerate poetry.lock by @muxator in https://github.com/bancaditalia/black-it/pull/9
+* Simplification of CIs by @AldoGl in https://github.com/bancaditalia/black-it/pull/13
+* CIs: add upload of codecoverage to Codecov by @AldoGl in https://github.com/bancaditalia/black-it/pull/15
+* Readme: add Codecov badge by @AldoGl in https://github.com/bancaditalia/black-it/pull/16
+* Improved documentation by @AldoGl in https://github.com/bancaditalia/black-it/pull/10
+* Update poetry.lock by @marcofavoritobi in https://github.com/bancaditalia/black-it/pull/18
+* Particle swarm sampler by @marcofavoritobi in https://github.com/bancaditalia/black-it/pull/17
+
+0.1.1 (2022-06-21)
+------------------
+* First release
