+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "black-it"
-version = "0.1.1"
+version = "0.1.2"
 description = "black-it: Black-box abm calibration kit"
 authors = [
     "Applied Research Team <appliedresearchteam@bancaditalia.it>",
