### 2022-11-02 #4622 @danielskatz
👋 @xuanxu - can you check on the error above for me?  I wonder if this is an issue with having multiple minipages within a figure, each with a label?  Otherwise, I don't see a problem in how the labels are created and used, but maybe I'm missing something?
