### 2022-11-03 5961a9e
Update paper-md to include JATS code for subfigures
