### 2022-11-03 #4622 @xuanxu
The origin of the error is that pandoc (the tool we use to produce pdf and metadata JATS files) can't handle subfigures like the fig1 and 2 in the paper.md. There's two options to solve this: One solution would be to combine both panels into a single image. Another solution is to include the jats code into the paper.md file so pandoc can use it.
I've [opened a pull request](https://github.com/bancaditalia/black-it/pull/21) that should solve the problem using the second option.
