### 2022-11-03 #4622 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
