+++ paper/paper.bib
@@ -129,7 +129,7 @@
 }
 
 @inproceedings{bajer2015benchmarking,
-  title={Benchmarking Gaussian processes and random forests surrogate models on the BBOB noiseless testbed},
+  title={Benchmarking {G}aussian processes and random forests surrogate models on the BBOB noiseless testbed},
   author={Bajer, Luk{\'a}{\v{s}} and Pitra, Zbyn{\v{e}}k and Hole{\v{n}}a, Martin},
   booktitle={Proceedings of the Companion Publication of the 2015 Annual Conference on Genetic and Evolutionary Computation},
   pages={1143--1150},
@@ -190,7 +190,7 @@
 }
 
 @article{farmer2022black,
-  title={Black-box Bayesian inference for economic agent-based models},
+  title={Black-box {B}ayesian inference for economic agent-based models},
   author={Dyer, Joel and Cannon, Patrick and Farmer, J Doyne and Schmon, Sebastian},
   journal={arXiv preprint arXiv:2202.00625},
   year={2022},
