### 2022-11-03 #4622 @danielskatz
👋 @AldoGl - I have some suggested small changes in https://github.com/bancaditalia/black-it/pull/22 - please merge this, or let me know what you disagree with, then we can proceed to acceptance
